Ported a few more color adjustment filters to the Mac. Tweaked the way that GPUImageView works on the Mac to support resizing correctly.

+++ framework/Source/GPUImageChromaKeyBlendFilter.m
@@ -99,14 +99,14 @@ NSString *const kGPUImageChromaKeyBlendFragmentShaderString = SHADER_STRING
     [self setVec3:colorToReplace forUniform:colorToReplaceUniform program:filterProgram];
 }
 
-- (void)setThresholdSensitivity:(CGFloat)newValue;
+- (void)setThresholdSensitivity:(GLfloat)newValue;
 {
     _thresholdSensitivity = newValue;
 
     [self setFloat:_thresholdSensitivity forUniform:thresholdSensitivityUniform program:filterProgram];
 }
 
-- (void)setSmoothing:(CGFloat)newValue;
+- (void)setSmoothing:(GLfloat)newValue;
 {
     _smoothing = newValue;
